fix: fuzz
